requirements.txt
# ZTI Pipeline Dependencies
# Core
httpx>=0.25.0
httpx-sse>=0.4.0
structlog>=23.0.0

# Database clients
python-arango>=8.0.0
qdrant-client>=1.6.0

# Scientific computing
numpy>=1.24.0
scikit-learn>=1.3.0
umap-learn>=0.5.0
hdbscan>=0.8.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
click>=8.0.0

